+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 HTTP client extension Change Log
 2.0.8 under development
 -----------------------
 
-- no changes in this release.
+- Enh #153: Allow configuring `JsonParser` to parse JSON as objects instead of arrays (CeBe)
 
 
 2.0.7 September 24, 2018